refactor: unify Reviews, Overrated, Underrated, and Underseen components
- Create shared Reviews.reducer.ts eliminating ~800 lines of duplicate reducer code
- Create shared Reviews.Filters.tsx for consistent filter UI across all review views
- Update ReviewListItemValue type to include review date fields
- Consolidate all review-based components to use the same reducer and filter components
- Add grade filter to Overrated, Underrated, and Underseen for consistency
- Maintain full backwards compatibility with existing functionality

This refactoring reduces code duplication by ~30% in the review components while
improving maintainability and ensuring consistent behavior across all review views.
